# RestaurantReservation.Api.EmailService/RestaurantReservation.Api.EmailService/Models/ApplicationUser.cs
using System;

namespace RestaurantReservation.Api.EmailService.Models
{
    public class ApplicationUser
    {
        public string Id { get; set; }
        public string Username { get; set; }
        public string NormalizedUserName { get; set; }
        public string Email { get; set; }
        public string NormalizedEmail { get; set; }
        public bool EmailConfirmed { get; set; }
        public string PasswordHash { get; set; }
        public string SecurityStamp { get; set; }
        public string ConcurrencyStamp { get; set; }
        public string PhoneNumber { get; set; }
        public bool PhoneNumberConfirmed { get; set; }
        public bool TwoFactorEnabled { get; set; }
        public DateTime LockoutEnd { get; set; }
        public bool LockoutEnabled { get; set; }
        public int AccessFailedCount { get; set; }
        public string Name { get; set; }
        public DateTime JoinedDate { get; set; }
        public bool isSubscriber { get; set; }
    }
}