# RestaurantReservation.Backend/RestaurantReservation.Api/bin/Debug/netcoreapp3.1/Swaggerv1.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Restaurant Reservation API v1",
    "description": "Restaurant Reservation API",
    "version": "v1"
  },
  "paths": {
...
    "/api/Authenticate/RefreshToken": {
      "post": {
        "tags": [
          "Authenticate"
        ],
        "summary": "Refresh User JWT Access Token When accessing the resources everytime.",
        "operationId": "RefreshToken",
        "requestBody": {
          "description": "Provides the user's current JWT access token and refresh token to compare database one",
          "content": {
            "application/json-patch+json; v=1.0": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            },
            "application/json; v=1.0": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            },
            "text/json; v=1.0": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            },
            "application/*+json; v=1.0": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Returns the user latest JWT access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserToken"
                }
              }
            }
          },
          "500": {
            "description": "Error Retrieving user refresh token from database",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Authenticate/ForgotPassword/{email}": {
      "get": {
        "tags": [
          "Authenticate"
        ],
        "summary": "Accepts an email input from user to send email with URL link for change password purposes.",
        "operationId": "ForgotPassword",
        "parameters": [
          {
            "name": "email",
            "in": "path",
            "description": "Email Address",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Email Address",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A Boolean response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          },
          "500": {
            "description": "Error sending email wiht URL to the user",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/api/Authenticate/ResetPassword": {
      "post": {
        "tags": [
          "Authenticate"
        ],
        "summary": "A combination of Reset Password Token + User's Email Address and Updated Password",
        "operationId": "ResetPassword",
        "requestBody": {
          "description": "ResetPasswordVM",
          "content": {
            "application/json-patch+json; v=1.0": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordVM"
              }
            },
            "application/json; v=1.0": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordVM"
              }
            },
            "text/json; v=1.0": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordVM"
              }
            },
            "application/*+json; v=1.0": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordVM"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "A Boolean response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          },
          "500": {
            "description": "Error updating user's account password",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApplicationUser": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "joinedDate": {
            "type": "string",
            "format": "date-time"
          },
          "role": {
            "type": "string",
            "nullable": true
          },
          "isSubscriber": {
            "type": "boolean"
          },
          "id": {
            "type": "string",
            "nullable": true
          },
          "userName": {
            "type": "string",
            "nullable": true
          },
          "normalizedUserName": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "normalizedEmail": {
            "type": "string",
            "nullable": true
          },
          "emailConfirmed": {
            "type": "boolean"
          },
          "passwordHash": {
            "type": "string",
            "nullable": true
          },
          "securityStamp": {
            "type": "string",
            "nullable": true
          },
          "concurrencyStamp": {
            "type": "string",
            "nullable": true
          },
          "phoneNumber": {
            "type": "string",
            "nullable": true
          },
          "phoneNumberConfirmed": {
            "type": "boolean"
          },
          "twoFactorEnabled": {
            "type": "boolean"
          },
          "lockoutEnd": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "lockoutEnabled": {
            "type": "boolean"
          },
          "accessFailedCount": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "RefreshRequest": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "nullable": true
          },
          "refreshToken": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "UserToken": {
        "type": "object",
        "properties": {
          "applicationUser": {
            "$ref": "#/components/schemas/ApplicationUser"
          },
          "token": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ResetPasswordVM": {
        "required": [
          "confirmPassword"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "nullable": true
          },
          "token": {
            "type": "string",
            "nullable": true
          },
          "password": {
            "type": "string",
            "format": "password",
            "nullable": true
          },
          "confirmPassword": {
            "type": "string",
            "format": "password"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "JWT Authorization header using the Bearer scheme. \r\n\r\n Enter 'Bearer' [space] and then your token in the text input below.\r\n\r\nExample: Bearer 12345abcdef",
        "name": "Authorization",
        "in": "header"
      }
    }
  }
}